Instrumentos/Codigos/deprecated/Q4/requirements.txt
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
idna==2.10
numpy==1.20.2
pandas==1.2.4
pycodestyle==2.7.0
python-dateutil==2.8.1
python-dotenv==0.17.0
pytz==2021.1
requests==2.25.1
six==1.15.0
toml==0.10.2
urllib3==1.26.4
